A large overhaul of the rfr functions was done. This was done because the previous functions were inefficient, especially when trying to parse the data by the semiconductor type or defect site. This new system has a function that does the random forest regression (the costly time step), and then a separate function that can break down predictions in different ways. This eliminates the need to do the costly step repetitively.
-------------------------------------------------
To get the predictions for each point by the RFR model, use 
rfr_predictor. This will return two dictionaries (train and test) of 
dataframes. The keys are the fold of cross validation in each 
dictionary. 

These dictionaries can then be plugged into rmse_calculator which can 
output tables of the RMSE calculated for the train and test sets at each 
fold of cross validation. rmse_calculator can also take inputs to break 
down how the model predicts by type of semiconductor or site of defect 
(change output_type to 'train' or 'site')

------------------------------------------------
For parity plots there are separate functions from the rmse 
calculations. 

To generate the prediction tables needed for the parity plot, use 
rfr_pp_predictor. This function does something similar to rfr_predictor, 
but it actually incorporates more rounds of cv so that multiple 
predictions can be made on the test points and uncertainty in 
predictions can be evaluted.

Then the function parity_plot can be used to visualize the predictions. 
The plot can be visualized highlighting different semiconductor types, 
or the different defect sites by changing output_type.